backend/pyproject.toml
[tool.poetry]
name = "keypro"
version = "0.1.0"
description = ""
authors = ["JorgeFMorais <jorgemorais1993@hotmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-rest-framework = "^0.1.0"
tzdata = "^2024.2"
psycopg2-binary = "^2.9.9"
djangorestframework-gis = "^1.1"
django-cors-headers = "^4.5.0"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
